Now using Java 21 to run Gradle
Spigot's latest 1.20 and 1.21 libraries require Gradle to run at least on Java 17. I have now switched to a Java 21 environment, so I needed to remove deprecated methods and constructors to avoid warnings when compiling.

All the code is still - and will always be - compiled for Java 8 (even if done by a Java 21 environment) for compatibility reasons.