Added compatibility option for scoreboard plugins
The new "settings.register-scoreboards" option in config.yml can be used to ensure compatibility - when disabled - with other plugins that set the player's scoreboard to a custom scoreboard and replace the main Vanilla one. Disabling this option is incompatible with the following features:
- scoreboard module
- custom suffix module
- rank sorting in tablist (on Sponge and Bukkit < 1.21.2)
- per-player translated names in tablist (on Bukkit)

The default setting's value changes based on the environment detected when adding it to config.yml.

Additionally, regarding the tablist:
- fixed a bug with HEARTS custom suffixes
- names are now updated once every TablistManager#getSendingTimeout() ms, just like headers and footers